admin/requirements.txt
django>=4.2
psycopg2-binary>=2.9
python-dotenv>=1.0
djangorestframework>=3.14
django-cors-headers>=4.0
whitenoise>=6.6
django-admin-interface>=0.28